리뷰 반영: 제네릭 제거, 조건부 프리페치, 타입 분리, 테스트 간소화
- usePaginatedSessions: 불필요한 제네릭 타입 파라미터 제거 (추론 활용)
- usePaginatedSessions: 마지막 페이지 이후 불필요한 프리페치 방지
- PaginatedResponse 타입을 api/types.ts로 분리
- 마지막 페이지 테스트를 1페이지=마지막 페이지 방식으로 간소화
