+++ packages/web/src/features/sessions/api/types.ts
@@ -1,0 +1,7 @@
+export interface PaginatedResponse<T> {
+  items: T[]
+  totalCount: number
+  page: number
+  pageSize: number
+  totalPages: number
+}
